+++ Escornabot/ButtonSetDigital.cpp
@@ -1,7 +1,7 @@
 // ButtonSetDigital.cpp
 /*
 
-Copyright (C) 2014-2017 Escornabot - http://escornabot.com
+Copyright (C) 2014-2018 Escornabot - http://escornabot.com
 
 This program is free software; you can redistribute it and/or
 modify it under the terms of the GNU General Public License