+++ Escornabot/Engine.h
@@ -1,7 +1,7 @@
 // Engine.h
 /*
 
-Copyright (C) 2014-2017 Escornabot - http://escornabot.com
+Copyright (C) 2014-2018 Escornabot - http://escornabot.com
 
 This program is free software; you can redistribute it and/or
 modify it under the terms of the GNU General Public License
@@ -48,11 +48,6 @@ public:
     virtual void init() = 0;
 
     /**
-     * Sets the default degrees value to use with turnRight() and turnLeft().
-     */
-    void setTurnDegrees(int16_t degrees) { _degrees = degrees; }
-
-    /**
      * Turns left or right an angle specified in degrees (from Escornabot's POV).
      * @param degrees Amount of degrees to turn. Positive is clockwise,
      *     negative is counter-clockwise.
@@ -60,21 +55,11 @@ public:
     virtual void turn(int16_t degrees) = 0;
 
     /**
-     * Turns right the configured amount of degrees.
-     */
-    void turnRight() { turn(_degrees); }
-
-    /**
-     * Turns left the configured amount of degrees.
-     */
-    void turnLeft() { turn(-_degrees); }
-
-    /**
      * Moves forward or backward direction (from Escornabot's POV).
-     * @param units Amount of units to move. Positive is forwards, negative
-     *     is backwards.
+     * @param advance_units Amount of units to move. Positive is forwards,
+     *      negative is backwards.
      */
-    virtual void moveStraight(int8_t units) = 0;
+    virtual void moveStraight(float advance_units) = 0;
 
     /**
      * Executes the movement program.
@@ -88,19 +73,27 @@ public:
 
     bool isExecuting() { return _program != NULL; }
 
+    bool isSquareDiagonals() { return _square_diagonals; }
+    void setSquareDiagonals(bool value)
+            { _square_diagonals = value; _current_degrees = 0; }
+
+    bool isAligned(int16_t degrees) { return _current_degrees % degrees == 0; }
 
 protected:
 
-    int16_t _degrees;
+    int16_t _current_degrees;
+    bool _square_diagonals;
 
     MoveList* _program;
     uint8_t _program_index;
     MOVE _getCurrentMove() { return _program->getMove(_program_index); }
 
     bool _is_cancelling;
-
     void _prepareMove();
 
+private:
+
+    float _calculateAdvanceUnits();
 };