+++ Escornabot/MoveList.cpp
@@ -1,7 +1,7 @@
 // MoveList.cpp
 /*
 
-Copyright (C) 2014-2017 Escornabot - http://escornabot.com
+Copyright (C) 2014-2018 Escornabot - http://escornabot.com
 
 This program is free software; you can redistribute it and/or
 modify it under the terms of the GNU General Public License
@@ -32,6 +32,8 @@ extern EventManager* EVENTS;
 MoveList::MoveList()
 {
     _move_count = 0;
+    _degrees = 90;
+    _alt_degrees = 45;
 }